Make draft card creation async to unblock dictionary response
The dictionary endpoint now returns immediately with the lookup result
while draft card creation (word normalization + translation AI calls)
runs asynchronously in a background thread.
