+++ server/src/main/java/io/github/mucsi96/learnlanguage/LearnLanguageApplication.java
@@ -2,8 +2,10 @@ package io.github.mucsi96.learnlanguage;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
+@EnableAsync
 public class LearnLanguageApplication {
 
 	public static void main(String[] args) {
