+++ server/src/main/java/io/github/mucsi96/learnlanguage/controller/DictionaryController.java
@@ -1,24 +1,15 @@
 package io.github.mucsi96.learnlanguage.controller;
 
-import java.util.Locale;
-
 import org.springframework.http.MediaType;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.github.mucsi96.learnlanguage.entity.Source;
-import io.github.mucsi96.learnlanguage.model.CardType;
 import io.github.mucsi96.learnlanguage.model.DictionaryRequest;
-import io.github.mucsi96.learnlanguage.model.LanguageLevel;
-import io.github.mucsi96.learnlanguage.model.SourceFormatType;
-import io.github.mucsi96.learnlanguage.model.SourceType;
 import io.github.mucsi96.learnlanguage.service.ApiTokenService;
 import io.github.mucsi96.learnlanguage.service.DictionaryService;
 import io.github.mucsi96.learnlanguage.service.DraftCardService;
-import io.github.mucsi96.learnlanguage.service.SourceService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
@@ -28,11 +19,9 @@ public class DictionaryController {
 
     private final ApiTokenService apiTokenService;
     private final DictionaryService dictionaryService;
-    private final SourceService sourceService;
     private final DraftCardService draftCardService;
 
     @PostMapping(value = "/dictionary", produces = MediaType.TEXT_PLAIN_VALUE)
-    @Transactional
     public String translate(
             @RequestHeader(value = "Authorization", required = false) String authorizationHeader,
             @Valid @RequestBody DictionaryRequest request) {
@@ -40,34 +29,10 @@ public class DictionaryController {
 
         if (request.getBookTitle() != null && request.getHighlightedWord() != null
                 && request.getSentence() != null) {
-            final Source source = getOrCreateSource(request.getBookTitle());
-            draftCardService.createDraftCard(source, request.getHighlightedWord(), request.getSentence());
+            draftCardService.createDraftCard(request.getBookTitle(), request.getHighlightedWord(),
+                    request.getSentence());
         }
 
         return dictionaryService.lookup(request);
-    }
-
-    private Source getOrCreateSource(String bookTitle) {
-        final String sourceId = toSourceId(bookTitle);
-
-        return sourceService.getSourceById(sourceId)
-                .orElseGet(() -> {
-                    final Source newSource = Source.builder()
-                            .id(sourceId)
-                            .name(bookTitle)
-                            .sourceType(SourceType.EBOOK_DICTIONARY)
-                            .startPage(1)
-                            .languageLevel(LanguageLevel.B1)
-                            .cardType(CardType.VOCABULARY)
-                            .formatType(SourceFormatType.WORD_LIST_WITH_EXAMPLES)
-                            .build();
-                    return sourceService.saveSource(newSource);
-                });
-    }
-
-    private static String toSourceId(String bookTitle) {
-        return bookTitle.toLowerCase(Locale.ROOT)
-                .replaceAll("[^a-z0-9]+", "-")
-                .replaceAll("^-|-$", "");
     }
 }
